+++ CHANGELOG.md
@@ -2,12 +2,18 @@
 
 All notable changes to this project will be documented in this file. This project uses [Semantic Versioning](https://semver.org/)
 
+## [Version 1.2.3](https://github.com/OpenWonderLabs/homebridge-switchbot-openapi/compare/v1.2.1...v1.2.2) (2021-01-12)
+
+### Changes
+
+- Added logging for IR Devices when connected to a [SwitchBot Hub Mini](https://www.switch-bot.com/products/switchbot-hub-mini) or [SwitchBot Hub Plus](https://www.switch-bot.com/products/switchbot-hub-plus).
+
 ## [Version 1.2.2](https://github.com/OpenWonderLabs/homebridge-switchbot-openapi/compare/v1.2.1...v1.2.2) (2021-01-12)
 
 ### Changes
 
-- Add Logging for [SwitchBot Hub Mini](https://www.switch-bot.com/products/switchbot-hub-mini), When discovered.
-- Add Logging for [SwitchBot Hub Plus](https://www.switch-bot.com/products/switchbot-hub-plus), When discovered.
+- Added logging for [SwitchBot Hub Mini](https://www.switch-bot.com/products/switchbot-hub-mini), When discovered.
+- Added logging for [SwitchBot Hub Plus](https://www.switch-bot.com/products/switchbot-hub-plus), When discovered.
 
 ## [Version 1.2.1](https://github.com/OpenWonderLabs/homebridge-switchbot-openapi/compare/v1.2.0...v1.2.1) (2021-01-11)